backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiomysql==0.2.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pyodbc==5.0.1
cx-Oracle==8.3.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Scheduler
apscheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
python-dotenv==1.0.0